requirements.txt
# requirements.txt
python-dotenv==1.0.0
requests-oauthlib==1.3.1
python-telegram-bot==20.3
apscheduler==3.10.1
cryptography==41.0.3